test: Introduce iroh-test with common logging infrastructure
This code was copy-pasted around too much.  This takes a simple and
clean approach to this by creating a new crate which is used as a
dev-dependency of all the other crates.

The upside is there is no messing about with cargo features and
delicate matching versions.

The main downside is that it can not provide testing utilities which
build on the code in the other crates.  Most notable this is currently
the testing derp server.
