+++ iroh-net/Cargo.toml
@@ -90,6 +90,7 @@ pretty_assertions = "1.4"
 rand_chacha = "0.3.1"
 tokio = { version = "1", features = ["io-util", "sync", "rt", "net", "fs", "macros", "time", "test-util"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+iroh-test = { path = "../iroh-test" }
 
 [build-dependencies]
 duct = "0.13.6"
