+++ iroh-net/src/magic_endpoint.rs
@@ -439,7 +439,7 @@ pub async fn get_peer_id(connection: &quinn::Connection) -> anyhow::Result<PeerI
 mod tests {
     use tracing::{info, info_span, Instrument};
 
-    use crate::test_utils::{run_derp_and_stun, setup_logging};
+    use crate::test_utils::run_derp_and_stun;
 
     use super::*;
 
@@ -448,7 +448,7 @@ mod tests {
     #[ignore]
     #[tokio::test]
     async fn magic_endpoint_connect_close() {
-        let _guard = setup_logging();
+        let _guard = iroh_test::logging::setup();
         let (derp_map, region_id, _guard) = run_derp_and_stun([127, 0, 0, 1].into()).await.unwrap();
         let server_keypair = Keypair::generate();
         let server_peer_id = PeerId::from(server_keypair.public());
