+++ iroh-net/src/magicsock.rs
@@ -2827,7 +2827,7 @@ pub(crate) mod tests {
         })
     }
 
-    pub fn setup_logging() {
+    pub fn setup_multithreaded_logging() {
         tracing_subscriber::registry()
             .with(tracing_subscriber::fmt::layer().with_writer(std::io::stderr))
             .with(EnvFilter::from_default_env())
@@ -2837,7 +2837,7 @@ pub(crate) mod tests {
 
     #[tokio::test(flavor = "multi_thread")]
     async fn test_two_devices_roundtrip_quinn_magic() -> Result<()> {
-        setup_logging();
+        setup_multithreaded_logging();
 
         let devices = Devices {
             stun_ip: "127.0.0.1".parse()?,
@@ -3009,7 +3009,7 @@ pub(crate) mod tests {
 
     #[tokio::test(flavor = "multi_thread")]
     async fn test_two_devices_setup_teardown() -> Result<()> {
-        setup_logging();
+        setup_multithreaded_logging();
         let devices = Devices {
             stun_ip: "127.0.0.1".parse()?,
         };
@@ -3053,7 +3053,7 @@ pub(crate) mod tests {
 
     #[tokio::test]
     async fn test_two_devices_roundtrip_quinn_raw() -> Result<()> {
-        setup_logging();
+        let _guard = iroh_test::logging::setup();
 
         let make_conn = |addr: SocketAddr| -> anyhow::Result<quinn::Endpoint> {
             let key = key::node::SecretKey::generate();
@@ -3197,7 +3197,7 @@ pub(crate) mod tests {
 
     #[tokio::test]
     async fn test_two_devices_roundtrip_quinn_rebinding_conn() -> Result<()> {
-        setup_logging();
+        let _guard = iroh_test::logging::setup();
 
         async fn make_conn(addr: SocketAddr) -> anyhow::Result<quinn::Endpoint> {
             let key = key::node::SecretKey::generate();
