+++ iroh-net/src/netcheck/reportgen/hairpin.rs
@@ -198,8 +198,6 @@ mod tests {
     use tokio::sync::mpsc::error::TrySendError;
     use tracing::info;
 
-    use crate::test_utils::setup_logging;
-
     use super::*;
 
     #[tokio::test]
@@ -213,7 +211,7 @@ mod tests {
     }
 
     async fn test_hairpin(hairpinning_works: bool) {
-        let _guard = setup_logging();
+        let _guard = iroh_test::logging::setup();
 
         // Setup fake netcheck and reportstate actors, hairpinning interacts with them.
         let (netcheck_tx, mut netcheck_rx) = mpsc::channel(32);
@@ -299,7 +297,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_client_drop() {
-        let _guard = setup_logging();
+        let _guard = iroh_test::logging::setup();
 
         // Setup fake netcheck and reportstate actors, hairpinning interacts with them.
         let (netcheck_tx, _netcheck_rx) = mpsc::channel(32);
