+++ iroh-test/README.md
@@ -1,0 +1,4 @@
+# iroh-test
+
+This crate contains testing support code for iroh.  It is only used
+for running the iroh test suite.
