+++ iroh-test/src/lib.rs
@@ -1,0 +1,3 @@
+//! Internal utilities to support testing.
+
+pub mod logging;
