+++ iroh/Cargo.toml
@@ -64,6 +64,7 @@ anyhow = { version = "1", features = ["backtrace"] }
 bytes = "1"
 duct = "0.13.6"
 genawaiter = { version = "0.99", features = ["futures03"] }
+iroh-test = { path = "../iroh-test" }
 nix = "0.26.2"
 postcard = "1"
 proptest = "1.2.0"
